+++ macos/Classes/InAppWebView/WebViewChannelDelegate.swift
@@ -845,11 +845,11 @@ public class WebViewChannelDelegate : ChannelDelegate {
     }
     
     public func onPermissionRequest(request: PermissionRequest, callback: PermissionRequestCallback) {
-        guard let channel = channel else {
+        if channel == nil {
             callback.defaultBehaviour(nil)
             return
         }
-        channel.invokeMethod("onPermissionRequest", arguments: request.toMap(), callback: callback)
+        channel?.invokeMethod("onPermissionRequest", arguments: request.toMap(), callback: callback)
     }
     
     public class ShouldOverrideUrlLoadingCallback : BaseCallbackResult<WKNavigationActionPolicy> {
@@ -869,11 +869,11 @@ public class WebViewChannelDelegate : ChannelDelegate {
     }
     
     public func shouldOverrideUrlLoading(navigationAction: WKNavigationAction, callback: ShouldOverrideUrlLoadingCallback) {
-        guard let channel = channel else {
+        if channel == nil {
             callback.defaultBehaviour(nil)
             return
         }
-        channel.invokeMethod("shouldOverrideUrlLoading", arguments: navigationAction.toMap(), callback: callback)
+        channel?.invokeMethod("shouldOverrideUrlLoading", arguments: navigationAction.toMap(), callback: callback)
     }
     
     public func onLoadStart(url: String?) {
@@ -924,7 +924,7 @@ public class WebViewChannelDelegate : ChannelDelegate {
     }
     
     public func onReceivedHttpAuthRequest(challenge: HttpAuthenticationChallenge, callback: ReceivedHttpAuthRequestCallback) {
-        guard let channel = channel else {
+        if channel == nil {
             callback.defaultBehaviour(nil)
             return
         }
@@ -932,8 +932,12 @@ public class WebViewChannelDelegate : ChannelDelegate {
         // https://github.com/pichillilorenzo/flutter_inappwebview/issues/1678
         DispatchQueue.global(qos: .background).async {
             let arguments = challenge.toMap()
-            DispatchQueue.main.async {
-                channel.invokeMethod("onReceivedHttpAuthRequest", arguments: arguments, callback: callback)
+            DispatchQueue.main.async { [weak self] in
+                if self?.channel == nil {
+                    callback.defaultBehaviour(nil)
+                    return
+                }
+                self?.channel?.invokeMethod("onReceivedHttpAuthRequest", arguments: arguments, callback: callback)
             }
         }
     }
@@ -952,7 +956,7 @@ public class WebViewChannelDelegate : ChannelDelegate {
     }
     
     public func onReceivedServerTrustAuthRequest(challenge: ServerTrustChallenge, callback: ReceivedServerTrustAuthRequestCallback) {
-        guard let channel = channel else {
+        if channel == nil {
             callback.defaultBehaviour(nil)
             return
         }
@@ -960,8 +964,12 @@ public class WebViewChannelDelegate : ChannelDelegate {
         // https://github.com/pichillilorenzo/flutter_inappwebview/issues/1678
         DispatchQueue.global(qos: .background).async {
             let arguments = challenge.toMap()
-            DispatchQueue.main.async {
-                channel.invokeMethod("onReceivedServerTrustAuthRequest", arguments: arguments, callback: callback)
+            DispatchQueue.main.async { [weak self] in
+                if self?.channel == nil {
+                    callback.defaultBehaviour(nil)
+                    return
+                }
+                self?.channel?.invokeMethod("onReceivedServerTrustAuthRequest", arguments: arguments, callback: callback)
             }
         }
     }
@@ -980,7 +988,7 @@ public class WebViewChannelDelegate : ChannelDelegate {
     }
     
     public func onReceivedClientCertRequest(challenge: ClientCertChallenge, callback: ReceivedClientCertRequestCallback) {
-        guard let channel = channel else {
+        if channel == nil {
             callback.defaultBehaviour(nil)
             return
         }
@@ -988,8 +996,12 @@ public class WebViewChannelDelegate : ChannelDelegate {
         // https://github.com/pichillilorenzo/flutter_inappwebview/issues/1678
         DispatchQueue.global(qos: .background).async {
             let arguments = challenge.toMap()
-            DispatchQueue.main.async {
-                channel.invokeMethod("onReceivedClientCertRequest", arguments: arguments, callback: callback)
+            DispatchQueue.main.async { [weak self] in
+                if self?.channel == nil {
+                    callback.defaultBehaviour(nil)
+                    return
+                }
+                self?.channel?.invokeMethod("onReceivedClientCertRequest", arguments: arguments, callback: callback)
             }
         }
     }
@@ -1019,12 +1031,12 @@ public class WebViewChannelDelegate : ChannelDelegate {
     }
     
     public func onLoadResourceWithCustomScheme(request: WebResourceRequest, callback: LoadResourceWithCustomSchemeCallback) {
-        guard let channel = channel else {
+        if channel == nil {
             callback.defaultBehaviour(nil)
             return
         }
         let arguments: [String: Any?] = ["request": request.toMap()]
-        channel.invokeMethod("onLoadResourceWithCustomScheme", arguments: arguments, callback: callback)
+        channel?.invokeMethod("onLoadResourceWithCustomScheme", arguments: arguments, callback: callback)
     }
     
     public class CallJsHandlerCallback : BaseCallbackResult<Any> {
@@ -1037,7 +1049,7 @@ public class WebViewChannelDelegate : ChannelDelegate {
     }
     
     public func onCallJsHandler(handlerName: String, args: String, callback: CallJsHandlerCallback) {
-        guard let channel = channel else {
+        if channel == nil {
             callback.defaultBehaviour(nil)
             return
         }
@@ -1045,7 +1057,7 @@ public class WebViewChannelDelegate : ChannelDelegate {
             "handlerName": handlerName,
             "args": args
         ]
-        channel.invokeMethod("onCallJsHandler", arguments: arguments, callback: callback)
+        channel?.invokeMethod("onCallJsHandler", arguments: arguments, callback: callback)
     }
     
     public class NavigationResponseCallback : BaseCallbackResult<WKNavigationResponsePolicy> {
@@ -1065,11 +1077,11 @@ public class WebViewChannelDelegate : ChannelDelegate {
     }
     
     public func onNavigationResponse(navigationResponse: WKNavigationResponse, callback: NavigationResponseCallback) {
-        guard let channel = channel else {
+        if channel == nil {
             callback.defaultBehaviour(nil)
             return
         }
-        channel.invokeMethod("onNavigationResponse", arguments: navigationResponse.toMap(), callback: callback)
+        channel?.invokeMethod("onNavigationResponse", arguments: navigationResponse.toMap(), callback: callback)
     }
     
     public class ShouldAllowDeprecatedTLSCallback : BaseCallbackResult<Bool> {
@@ -1089,7 +1101,7 @@ public class WebViewChannelDelegate : ChannelDelegate {
     }
     
     public func shouldAllowDeprecatedTLS(challenge: URLAuthenticationChallenge, callback: ShouldAllowDeprecatedTLSCallback) {
-        guard let channel = channel else {
+        if channel == nil {
             callback.defaultBehaviour(nil)
             return
         }
@@ -1097,8 +1109,12 @@ public class WebViewChannelDelegate : ChannelDelegate {
         // https://github.com/pichillilorenzo/flutter_inappwebview/issues/1678
         DispatchQueue.global(qos: .background).async {
             let arguments = challenge.toMap()
-            DispatchQueue.main.async {
-                channel.invokeMethod("shouldAllowDeprecatedTLS", arguments: arguments, callback: callback)
+            DispatchQueue.main.async { [weak self] in
+                if self?.channel == nil {
+                    callback.defaultBehaviour(nil)
+                    return
+                }
+                self?.channel?.invokeMethod("shouldAllowDeprecatedTLS", arguments: arguments, callback: callback)
             }
         }
     }
@@ -1141,7 +1157,7 @@ public class WebViewChannelDelegate : ChannelDelegate {
     }
     
     public func onPrintRequest(url: URL?, printJobId: String?, callback: PrintRequestCallback) {
-        guard let channel = channel else {
+        if channel == nil {
             callback.defaultBehaviour(nil)
             return
         }
@@ -1149,7 +1165,7 @@ public class WebViewChannelDelegate : ChannelDelegate {
             "url": url?.absoluteString,
             "printJobId": printJobId,
         ]
-        channel.invokeMethod("onPrintRequest", arguments: arguments, callback: callback)
+        channel?.invokeMethod("onPrintRequest", arguments: arguments, callback: callback)
     }
     
     public override func dispose() {